EVALUE-406 Flash report new design (#158)

+++ src/external_api/schemas/flash_report.py
@@ -9,7 +9,7 @@ class FlashReportFormType(BaseModel):
     model: str
     type: str
     version: str | None = None
-    odometer: float
+    odometer: int
     email: EmailStr
     language: LanguageEnum
